requirements.txt
requests
python-dotenv
yt-dlp
openai